+++ tests/test_app/TestApp/Test/TestCase/Model/Table/PostsTableTest.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Test\TestCase\Model\Table;
+
+use MeTools\TestSuite\TestCase;
+class PostsTableTest extends TestCase
+{
+}